verification/requirements.txt
pytest
pytest-playwright
requests
pixelmatch
Pillow
